add tcp template server
